refactor: Refactored schemas
