+++ src/app/api/schemas/__init__.py
@@ -1,0 +1,12 @@
+from .base import *
+from .accesses import *
+from .alerts import *
+from .devices import *
+from .events import *
+from .groups import *
+from .installations import *
+from .login import *
+from .media import *
+from .sites import *
+from .users import *
+from .webhooks import *
